[MISC] Move seqan::views3::translate and seqan::views::translate_single to seqan3/alphabet/views/
