+++ include/seqan3/alphabet/views/translate_join.hpp
@@ -17,10 +17,10 @@
 #include <vector>
 #include <stdexcept>
 
+#include <seqan3/alphabet/views/translate.hpp>
 #include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/range/container/small_string.hpp>
 #include <seqan3/range/detail/random_access_iterator.hpp>
-#include <seqan3/range/views/translate.hpp>
 
 namespace seqan3::detail
 {
