+++ include/seqan3/range/views/all.hpp
@@ -17,6 +17,7 @@
 #include <seqan3/alphabet/views/rank_to.hpp>
 #include <seqan3/alphabet/views/to_char.hpp>
 #include <seqan3/alphabet/views/to_rank.hpp>
+#include <seqan3/alphabet/views/translate.hpp>
 #include <seqan3/alphabet/views/trim_quality.hpp>
 #include <seqan3/range/views/get.hpp>
 #include <seqan3/range/views/interleave.hpp>
@@ -28,7 +29,6 @@
 #include <seqan3/range/views/take_line.hpp>
 #include <seqan3/range/views/take_until.hpp>
 #include <seqan3/range/views/to.hpp>
-#include <seqan3/range/views/translate.hpp>
 #include <seqan3/utility/views/convert.hpp>
 #include <seqan3/utility/views/enforce_random_access.hpp>
 #include <seqan3/utility/views/deep.hpp>
