+++ test/performance/range/views/view_translate_1D_benchmark.cpp
@@ -11,8 +11,8 @@
 #include <benchmark/benchmark.h>
 
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
+#include <seqan3/alphabet/views/translate.hpp>
 #include <seqan3/range/views/to.hpp>
-#include <seqan3/range/views/translate.hpp>
 #include <seqan3/test/performance/sequence_generator.hpp>
 #include <seqan3/test/seqan2.hpp>
