+++ test/snippet/range/views/translate_dna5.cpp
@@ -1,7 +1,7 @@
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
 #include <seqan3/alphabet/views/complement.hpp>
+#include <seqan3/alphabet/views/translate.hpp>
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/range/views/translate.hpp>
 
 int main()
 {
