+++ test/snippet/range/views/translate_usage.cpp
@@ -1,6 +1,6 @@
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
 #include <seqan3/alphabet/views/complement.hpp>
-#include <seqan3/range/views/translate.hpp>
+#include <seqan3/alphabet/views/translate.hpp>
 
 int main()
 {
